requirements.txt
Django==3.2.4
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.9.2
psycopg2==2.9.1
